try compaction on batch completion
